requirements.txt
APScheduler>=3.10.4
feedparser>=6.0.11
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.32.3
google-api-python-client>=2.137.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
python-dateutil>=2.9.0.post0
PyYAML>=6.0.2
tenacity>=9.0.0
